requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.5.2
beautifulsoup4==4.11.1
blessed==1.19.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-decouple==2.1
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jinxed==1.2.0
lxml==4.9.2
numpy==1.23.5
oracledb==1.2.2
pandas==1.5.2
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
redis==3.5.3
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
wcwidth==0.2.5
#pysftp
paramiko
psycopg2-binary
#pymqi
gunicorn
apscheduler